refactor(test): Move e2e tests to dedicated e2e/ directory
Improves test organization by separating end-to-end tests from the
main package. Changes package from 'main' to 'e2e' since these tests
use httptest and don't need main package access.

- e2e_test.go → e2e/scenario_test.go
- e2e_cpu_host_test.go → e2e/cpu_host_test.go
